# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Documentation สำหรับเกม Headliner",
        "title": "Headliner API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/achievements/": {
            "get": {
                "description": "ดึงรายการเงื่อนไขความสำเร็จ (Achievement) ทั้งหมดที่มีในระบบ",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Achievement"
                ],
                "summary": "ดึงข้อมูล Achievement ทั้งหมด",
                "responses": {
                    "200": {
                        "description": "ดึงข้อมูลสำเร็จ",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "เซิร์ฟเวอร์มีปัญหา",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "เพิ่มข้อมูลความสำเร็จ (Achievement) ใหม่เข้าสู่ระบบ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Achievement"
                ],
                "summary": "สร้าง Achievement ใหม่",
                "parameters": [
                    {
                        "description": "ข้อมูล Achievement ที่ต้องการสร้าง",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Entities.Achievement"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "สร้างสำเร็จ",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "ข้อมูลไม่ถูกต้อง",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "เซิร์ฟเวอร์มีปัญหา",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/achievements/unlock": {
            "post": {
                "description": "บันทึกข้อมูลว่าผู้เล่นได้รับ (Unlock) Achievement นี้เรียบร้อยแล้ว",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Achievement"
                ],
                "summary": "ปลดล็อก Achievement ให้ผู้เล่น",
                "parameters": [
                    {
                        "description": "ข้อมูล User ID และ Achievement ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AchievementAdapter.UnlockAchievementRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ปลดล็อกสำเร็จ",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "ข้อมูลไม่ถูกต้อง",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "เซิร์ฟเวอร์มีปัญหา",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/achievements/user/{userId}": {
            "get": {
                "description": "ดูรายการว่าผู้เล่นคนนี้ทำเงื่อนไข Achievement อะไรสำเร็จไปแล้วบ้าง",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Achievement"
                ],
                "summary": "ดึงรายการ Achievement ที่ผู้เล่นปลดล็อกแล้ว",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "รหัส User ID ของผู้เล่น",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ดึงข้อมูลสำเร็จ",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "รูปแบบ User ID ไม่ถูกต้อง",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "เซิร์ฟเวอร์มีปัญหา",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/achievements/{id}": {
            "get": {
                "description": "ค้นหาและดึงรายละเอียดของ Achievement แบบเจาะจงด้วย ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Achievement"
                ],
                "summary": "ดึงข้อมูล Achievement ตาม ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "รหัส Achievement ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ดึงข้อมูลสำเร็จ",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "รูปแบบ ID ไม่ถูกต้อง",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "ไม่พบข้อมูล",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "เซิร์ฟเวอร์มีปัญหา",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "AchievementAdapter.UnlockAchievementRequest": {
            "type": "object",
            "properties": {
                "achievement_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "Entities.Achievement": {
            "type": "object"
        },
...
    }
}